refactor: trocear NotificationsBottomSheet en components/notifications (Fase 1.9)
938 → 365 líneas. Extraídos a components/notifications/:
- NotificationDetail.tsx (vista en grande)
- NotificationListItem.tsx (tarjeta de la lista, swipe + chips; lógica por props)
- notificationDisplay.ts (helpers puros normalizeRoute/getRouteLabel/formatDate)
  con test (__tests__/notificationDisplay.test.ts, 14 tests)

Cero cambios de comportamiento (acción del chip movida a handler con la misma
lógica; quitados 2 estilos sin uso). Solo app/(tabs)/index.tsx consume el sheet
y su import no cambia. typecheck/typecheck:tests/lint(0 err)/test(197) ok.
